Import symbols from sickchill.helper.common instead of the package, move tests for the moved methods Make pymediainfo optional
