+++ pyproject.toml
@@ -216,7 +216,7 @@ validators = ">=0.18.2,<0.21.0"
 enzyme = "^0.4.1"
 python3-fanart = "^2.0.0"
 gntp = "^1.0.3"
-pymediainfo = "^5.0.3"
+pymediainfo = {version = "^5.0.3", optional = true}
 "tus.py" = "^1.3.4"
 win-inet-pton = "^1.1.0"
 PyNMA = "^1.0"
