+++ sickchill/tv.py
@@ -16,11 +16,10 @@ from unidecode import unidecode
 from urllib3.exceptions import MaxRetryError, NewConnectionError
 
 import sickchill
-import sickchill.helper.common
 import sickchill.oldbeard.providers
 import sickchill.oldbeard.scene_numbering
 from sickchill import logger, settings
-from sickchill.helper.common import dateTimeFormat, episode_num, remove_extension, replace_extension, sanitize_filename, try_int
+from sickchill.helper.common import dateTimeFormat, episode_num, is_media_file, remove_extension, replace_extension, sanitize_filename, try_int
 from sickchill.helper.exceptions import (
     EpisodeDeletedException,
     EpisodeNotFoundException,
@@ -664,7 +663,7 @@ class TVShow(object):
                     curEp.status = Quality.compositeStatus(DOWNLOADED, newQuality)
 
             # check for status/quality changes as long as it's a new file
-            elif not same_file and sickchill.helper.common.is_media_file(filepath) and curEp.status not in Quality.DOWNLOADED + Quality.ARCHIVED + [IGNORED]:
+            elif not same_file and is_media_file(filepath) and curEp.status not in Quality.DOWNLOADED + Quality.ARCHIVED + [IGNORED]:
                 oldStatus, oldQuality = Quality.splitCompositeStatus(curEp.status)
                 newQuality = Quality.nameQuality(filepath, self.is_anime)
 
@@ -1727,7 +1726,7 @@ class TVEpisode(object):
                 logger.debug("Not touching status [ {0} ] It could be skipped/ignored/snatched/archived".format(statusStrings[self.status]))
 
         # if we have a media file then it's downloaded
-        elif sickchill.helper.common.is_media_file(self.location):
+        elif is_media_file(self.location):
             # leave propers alone, you have to either post-process them or manually change them back
             if self.status not in Quality.SNATCHED_PROPER + Quality.DOWNLOADED + Quality.SNATCHED + Quality.ARCHIVED:
                 logger.debug(f"5 Status changes from {self.status} to {Quality.statusFromName(self.location)}")
@@ -1750,7 +1749,7 @@ class TVEpisode(object):
 
         if self.location != "":
 
-            if self.status == UNKNOWN and sickchill.helper.common.is_media_file(self.location):
+            if self.status == UNKNOWN and is_media_file(self.location):
                 logger.debug(f"7 Status changes from {self.status} to {Quality.statusFromName(self.location, anime=self.show.is_anime)}")
                 self.status = Quality.statusFromName(self.location, anime=self.show.is_anime)
